requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
Faker==22.5.1
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gradio==4.14.0
gradio_client==0.8.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mysql-connector-python==8.3.0
nest-asyncio==1.5.9
numpy==1.26.3
openai==1.8.0
orjson==3.9.10
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
stack-data==0.6.3
starlette==0.35.1
tenacity==8.2.3
tomlkit==0.12.0
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
wcwidth==0.2.13
websockets==11.0.3
